reformatted data generation
