+++ recommender/validation.py
@@ -8,6 +8,9 @@ import pandas as pd
 from rich.console import Console
 from rich.table import Table
 
+import sys
+  
+sys.path.insert(0, '../dataset_generator/')
 from dataset import get_data
