+++ requirements.txt
@@ -1,0 +1,78 @@
+altair==4.1.0
+appnope==0.1.2
+argon2-cffi==21.1.0
+attrs==21.2.0
+backcall==0.2.0
+bleach==4.1.0
+certifi==2021.5.30
+cffi==1.14.6
+charset-normalizer==2.0.6
+colorama==0.4.4
+commonmark==0.9.1
+debugpy==1.4.3
+decorator==5.1.0
+defusedxml==0.7.1
+entrypoints==0.3
+idna==3.2
+ipykernel==6.4.1
+ipython==7.28.0
+ipython-genutils==0.2.0
+ipywidgets==7.6.5
+jedi==0.18.0
+Jinja2==3.0.1
+joblib==1.0.1
+jsonschema==4.0.1
+jupyter==1.0.0
+jupyter-client==7.0.5
+jupyter-console==6.4.0
+jupyter-core==4.8.1
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.2
+lightfm==1.16
+MarkupSafe==2.0.1
+matplotlib-inline==0.1.3
+mistune==0.8.4
+nbclient==0.5.4
+nbconvert==6.2.0
+nbformat==5.1.3
+nest-asyncio==1.5.1
+networkx==2.6.3
+notebook==6.4.4
+numpy==1.21.2
+packaging==21.0
+pandas==1.3.3
+pandocfilters==1.5.0
+parso==0.8.2
+pexpect==4.8.0
+pickleshare==0.7.5
+prometheus-client==0.11.0
+prompt-toolkit==3.0.20
+ptyprocess==0.7.0
+pycparser==2.20
+Pygments==2.10.0
+pyparsing==2.4.7
+pyrsistent==0.18.0
+python-dateutil==2.8.2
+pytz==2021.3
+pyzmq==22.3.0
+qtconsole==5.1.1
+QtPy==1.11.2
+requests==2.26.0
+rich==10.11.0
+scikit-learn==1.0
+scipy==1.7.1
+Send2Trash==1.8.0
+six==1.16.0
+tabulate==0.8.9
+terminado==0.12.1
+testpath==0.5.0
+threadpoolctl==3.0.0
+toolz==0.11.1
+tornado==6.1
+tqdm==4.62.3
+traitlets==5.1.0
+urllib3==1.26.7
+vega==3.5.0
+wcwidth==0.2.5
+webencodings==0.5.1
+widgetsnbextension==3.5.1
